+++ NEWS.md
@@ -1,7 +1,8 @@
 # camtrapR 2.1.0
 
-## new function
+## new functions
 * OCRdataFields: allows automatic reading of text from data fields in images via optical character recognition (OCR), requires magick and tesseract packages. 
+* writeDateTimeOriginal: Write values to DateTimeOriginal tag to image metadata
 
 ## new features
 * cameraOperation: setup/ retrieval / problem columns can be interpreted as date-time format, then computes effort as fraction of day. (Thank you to Zoé Siegel for the suggestion, Fridolin Zimmermann and the Department of Ecology & Evolution from the University of Lausanne for funding the development of the feature)